Abstract event-emitter into separate class
